+++ embedded-can/src/lib.rs
@@ -61,6 +61,14 @@ pub trait Error: core::fmt::Debug {
     fn kind(&self) -> ErrorKind;
 }
 
+impl core::fmt::Display for dyn Error {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+        write!(f, "{:?}", self)
+    }
+}
+
+impl core::error::Error for dyn Error {}
+
 impl Error for core::convert::Infallible {
     fn kind(&self) -> ErrorKind {
         match *self {}