+++ embedded-io/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- Added blanket `core::error::Error` and `core::fmt::Display` implementations for the custom `Error` traits
+- Migrated `std` feature-gated `std::error::Error` implementations to `core::error::Error`
+- Increased MSRV to 1.81 due to `core::error::Error`
+
+
 ## 0.6.1 - 2023-10-22
 
 - Make `SliceWriteError` publicly available.