+++ embedded-io/CHANGELOG.md
@@ -5,10 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 0.7.0 - unreleased
 
-- Added `core::error::Error` implementations for every custom `impl Error`
-- Migrated `std` feature-gated `std::error::Error` implementations to `core::error::Error`
+- Added trait dependency on `core::error::Error` to this crate's `Error` trait
+  - Implemented the trait on all provided implementations
+  - Migrated `std` feature-gated `std::error::Error` implementations to `core::error::Error`
+- Updated `defmt` dependency to 1.0; feature renamed from `defmt_03` to `defmt`
 - Increased MSRV to 1.81 due to `core::error::Error`
 - Implemented `ReadReady` for `&[u8]` and `WriteReady` for `&mut [u8]`